requirements.txt
aiosignal==1.3.1
attrs==24.2.0
boto3==1.35.2
botocore==1.35.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dotenv-python==0.0.1
filelock==3.15.4
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
msgpack==1.0.8
numpy==1.26.4
packaging==24.1
pandas==2.2.2
protobuf==5.27.3
psycopg2-binary==2.9.9
pyarrow==17.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
ray==2.34.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.2
setuptools==73.0.1
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wheel==0.44.0